// Angular/12-12/site/src/app/models/cadastro-data.model.ts
export interface CadastroData {
    name: string;
    cpf: string;
    birthdate: string;
    zipCode: string;
    street: string;
    number: number;
    complement: string;
    district: string;
    state: string;
    city: string;
    email: string;
    phone: string;
    username: string;
    password: string;
  }